Old tests removed. Now they are in sync with RS sirepo. Need to create new ones.
