--- tests/sirepo_parser_data/t2.json
@@ -1,135 +1,0 @@
-{
-    "models": {
-        "beamline": [],
-        "electronBeam": {
-            "beamSelector": "NSLS-II Low Beta Final",
-            "current": 0.5,
-            "energy": 3.0,
-            "energyDeviation": 0.0,
-            "horizontalAlpha": 0.0,
-            "horizontalBeta": 2.02,
-            "horizontalDispersion": 0.0,
-            "horizontalDispersionDerivative": 0.0,
-            "horizontalEmittance": 0.5499999999999999,
-            "horizontalPosition": 0.0,
-            "isReadOnly": true,
-            "name": "NSLS-II Low Beta Final",
-            "rmsSpread": 0.00089,
-            "verticalAlpha": 0.0,
-            "verticalBeta": 1.06,
-            "verticalDispersion": 0.0,
-            "verticalDispersionDerivative": 0.0,
-            "verticalEmittance": 0.008,
-            "verticalPosition": 0.0
-        },
-        "electronBeams": [],
-        "fluxReport": {
-            "azimuthalPrecision": 1.0,
-            "distanceFromSource": 20,
-            "finalEnergy": 20000.0,
-            "fluxType": 1,
-            "horizontalApertureSize": 1.0,
-            "horizontalPosition": 0.0,
-            "initialEnergy": 100.0,
-            "longitudinalPrecision": 1.0,
-            "photonEnergyPointCount": 10000,
-            "polarization": 6,
-            "verticalApertureSize": 1.0,
-            "verticalPosition": 0.0
-        },
-        "gaussianBeam": {
-            "energyPerPulse": 0.001,
-            "polarization": 1,
-            "rmsPulseDuration": 0.1,
-            "rmsSizeX": 9.787230000000001,
-            "rmsSizeY": 9.787230000000001,
-            "waistAngleX": 0.0,
-            "waistAngleY": 0.0,
-            "waistX": 0.0,
-            "waistY": 0.0,
-            "waistZ": 0.0
-        },
-        "initialIntensityReport": {
-            "characteristic": 0,
-            "horizontalPosition": 0.0,
-            "horizontalRange": 0.4,
-            "polarization": 6,
-            "precision": 0.01,
-            "verticalPosition": 0.0,
-            "verticalRange": 0.6
-        },
-        "intensityReport": {
-            "distanceFromSource": 20,
-            "finalEnergy": 20000.0,
-            "horizontalPosition": 0.0,
-            "initialEnergy": 100.0,
-            "photonEnergyPointCount": 10000,
-            "polarization": 6,
-            "precision": 0.01,
-            "verticalPosition": 0.0
-        },
-        "multipole": {
-            "distribution": "n",
-            "field": 0.4,
-            "length": 3.0,
-            "order": 1
-        },
-        "postPropagation": [],
-        "powerDensityReport": {
-            "distanceFromSource": 20,
-            "horizontalPointCount": 100,
-            "horizontalPosition": 0.0,
-            "horizontalRange": 15.0,
-            "method": 1,
-            "precision": 1.0,
-            "verticalPointCount": 100,
-            "verticalPosition": 0.0,
-            "verticalRange": 15.0
-        },
-        "propagation": {},
-        "simulation": {
-            "facility": "Undulator",
-            "horizontalPointCount": 100,
-            "isExample": 0,
-            "name": "Undulator Radiation",
-            "photonEnergy": 9000.0,
-            "sampleFactor": 1.0,
-            "simulationId": null,
-            "sourceType": "u",
-            "verticalPointCount": 100
-        },
-        "sourceIntensityReport": {
-            "characteristic": 0,
-            "distanceFromSource": 20,
-            "horizontalPosition": 0,
-            "horizontalRange": 0.5,
-            "polarization": 6,
-            "sampleFactor": 0,
-            "verticalPosition": 0,
-            "verticalRange": 0.5
-        },
-        "undulator": {
-            "horizontalAmplitude": 0.0,
-            "horizontalInitialPhase": 0.0,
-            "horizontalSymmetry": 1,
-            "length": 3.0,
-            "longitudinalPosition": 0.0,
-            "period": 20.0,
-            "verticalAmplitude": 0.88770981,
-            "verticalInitialPhase": 0.0,
-            "verticalSymmetry": -1
-        },
-        "watchpointReport": {
-            "characteristic": 0,
-            "horizontalPosition": 0.0,
-            "horizontalRange": 0.4,
-            "polarization": 6,
-            "precision": 0.01,
-            "verticalPosition": 0.0,
-            "verticalRange": 0.6
-        }
-    },
-    "report": "",
-    "simulationType": "srw",
-    "version": "20160202"
-}
