+++ tests/sirepo_parser_test.py
@@ -6,15 +6,18 @@ u"""PyTest for :mod:`sirepo.sirepo_parser`
 """
 from __future__ import absolute_import, division, print_function
 
-from pykern import pkio
+import pytest
+
+import py
+
 from pykern import pkunit
+from pykern import pkio
 
 from sirepo.sirepo_parser import main
 
-
 def test_sirepo_parser():
     with pkunit.save_chdir_work():
-        for b in ('t1', ): # 't2', 't3',):
+        for b in ('t1',):
             in_py = str(pkunit.data_dir().join('{}.py'.format(b)))
             main(in_py, debug=False)
             actual_fn = 'parsed_sirepo.json'
